[bazel] Added manual workflow to publish a tag to BCR (#192)
Adds the publish.yaml workflow to manually release a tag to the Bazel Central Registry (BCR). This workflow uses the official reusable publish-to-bcr workflow from the bazel team.

BCR template files are added to mirror the existing files in the latest manual bcr release for gz-utils.

Also added the compatibility_level field in MODULE.bazel to mirror what is set in BCR for the latest release.

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>